carrier table and edit 1

+++ tables/carrier_edit.js
@@ -1,13 +1,18 @@
-const cached_carrier_Data = localStorage.getItem('data');
+const cached_carrier_id = localStorage.getItem('carrier_id');
 
-if (cached_carrier_Data) {
-    let felbontva = cached_carrier_Data.split(",");
-
-    document.getElementById('input_carrier_ekaer_id').value = felbontva[0];
-    document.getElementById('input_carrier_name').value = felbontva[1];
-    document.getElementById('input_carrier_memo').value = felbontva[2];
-    document.getElementById('input_carrier_id').value = felbontva[3];
-   
+if (cached_carrier_id) {
+    fetch('http://localhost:3001/carriers/'+ cached_carrier_id)
+    .then(response => response.json())
+    .then( datas => {
+        // console.log(datas);
+        datas.map( data => {
+            // console.log(data);
+            document.getElementById('name').value = data.name;
+            document.getElementById('ekaer_id').value = data.ekaer_id;
+            document.getElementById('memo').value = data.memo;
+            document.getElementById('id').value = data.id;
+        })
+    })
     localStorage.clear();
 }
 
@@ -15,40 +20,38 @@ const submit_button_carrier = document.getElementById("submit_button");
 
 submit_button_carrier.addEventListener("click", (event) => {
     event.preventDefault();
-    create_and_update_user();
+    create_and_update_carrier();
 })
 
-function create_and_update_user() {
- var input_carrier_ekaer_id = document.getElementById('input_carrier_ekaer_id').value;
- var input_carrier_name = document.getElementById('input_carrier_name').value;
- var input_carrier_memo = document.getElementById('input_carrier_memo').value;
- var input_carrier_id = document.getElementById('input_carrier_id').value;
+function create_and_update_carrier() {
+    var name = document.getElementById('name').value;
+    var ekaer_id = document.getElementById('ekaer_id').value;
+    var memo = document.getElementById('memo').value;
+    var id = document.getElementById('id').value;
  
     let data_to_send = {
-        "id":  input_carrier_id,
-        "ekaer_id": input_carrier_ekaer_id,
-        "name": input_carrier_name,
-        "memo": input_carrier_memo,
-         
+        "id":  id,
+        "name": name,
+        "ekaer_id": ekaer_id,
+        "memo": memo,
     }
 
-    if (input_carrier_id === '') {
-        fetch("http://localhost:3001/carriers", {
-            method: "POST",
-            body: JSON.stringify(data_to_send),
-            headers: {
-                "Content-type": "application/json"
-            }
-        }).then( res => console.log( res))
-    } else {
-        fetch(`http://localhost:3001/carriers/${input_carrier_id}`, {
-            method: "PUT",
-            body: JSON.stringify(data_to_send),
-            headers: {
-                "Content-type": "application/json"
-            }
-        }).then( res => console.log( res))
-    }
+    console.log( 'data?to?send ', data_to_send);
+    
+    let amethod;
+    if (id === '') amethod = "POST"
+    else amethod= "PUT";
+    fetch("http://localhost:3001/carriers", {
+        method: amethod,
+        body: JSON.stringify(data_to_send),
+        headers: {
+            "Content-type": "application/json"
+        }
+    }).then( res => {
+        console.log( res)
+        localStorage.setItem('back_id', id);
+        redirectToCarrierTable();
+    })
 }
 
 
@@ -56,23 +59,20 @@ const delete_carrier_button = document.getElementById("delete_button");
 
 delete_carrier_button.addEventListener("click", (event) => {
     event.preventDefault();
-    delete_user();
+    delete_carrier();
 })
 
+function delete_carrier() {
+    var id = document.getElementById('id').value;
 
-
-
-function delete_user() {
-    var input_carrier_id = document.getElementById('input_carrier_id').value;
-
-    fetch(`http://localhost:3001/carriers/${input_carrier_id}`, {
-        method: "DELETE"
-    })
- 
-      setTimeout(redirectToUserTable, 300);
-
-   
-   
+    if ( confirm( 'Biztos benne? Töröljük a szállítmányozót? '+ document.getElementById('name').value)) {
+        fetch(`http://localhost:3001/carriers/${id}`, {
+            method: "DELETE"
+        }).then( result => {
+            console.log( result);
+            redirectToUserTable();
+        })
+    }
 }
 
 function redirectToUserTable() {
